+++ .flake8
@@ -1,5 +1,5 @@
 [flake8]
-select = BLK,C,E,F,I,W
+select = BLK,C,E,F,W
 ignore = E203,W503
 max-line-length = 88
 max-complexity = 10
