+++ .isort.cfg
@@ -1,5 +1,5 @@
 [settings]
 use_parentheses=true
-known_first_party=limette
+known_first_party=citric
 multi_line_output=3
 include_trailing_comma=true
