+++ HISTORY.rst
@@ -1,47 +1,2 @@
 History
 =======
-
-1.0.3 (xxxx-xx-xx)
-------------------
-
-- Raise `LimeSurveyError` when JSON-RPC is not enabled.
-
-.. _section:
-
-1.0.2 (2019-11-23)
-------------------
-
--  Change Content-Type of project description.
--  Fix flake8 check by using a raw dosctring.
--  Use requests-mock pre-packaged pytest fixture.
-
-.. _section:
-
-1.0.1 (2019-11-23)
-------------------
-
--  New badges: supported python runtimes, codecov, pyup.io Python 3
--  Render docs in reStructuredText using Sphinx
--  Upgrade ``pip`` to 19.3.1
-
-.. _section:
-
-1.0.0 (2019-11-23)
-------------------
-
--  Pin dependencies to their latest versions.
--  limette now requires Python 3.6 or later.
-
-.. _section:
-
-0.0.2 (2019-11-23)
-------------------
-
--  First release on PyPI.
-
-.. _section:
-
-0.0.1 (2019-11-22)
-------------------
-
-- Low-level RPC API
