+++ src/citric/exceptions.py
