+++ tests/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-"""Unit test package for limette."""
+"""Unit test package for citric."""
